SessionHelper, secured controllers, comments

+++ app/models/ColorSource.java
@@ -8,7 +8,6 @@ import javax.persistence.Id;
 import javax.persistence.Inheritance;
 import javax.persistence.InheritanceType;
 import javax.persistence.Table;
-import java.math.BigInteger;
 import java.net.URI;
 
 /** {@code ColorSource} represents an entity defining the color name and/or the color itself. */
@@ -20,7 +19,7 @@ public class ColorSource {
   /** The id of this {@link ColorSource}. */
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private BigInteger id;
+  private Long id;
   /** The name. */
   @Column(name = "name", nullable = false)
   private String name;
@@ -29,7 +28,7 @@ public class ColorSource {
   private URI url;
 
   /** @return the id of this {@link ColorSource}. */
-  public BigInteger getId() { return id; }
+  public Long getId() { return id; }
 
   /** @return the name of the source. */
   public String getName() { return name; }
